requirements.txt
streamlit
langchain
openai
faiss-cpu
tiktoken
python-dotenv
